// src/types/generated/contract.ts
export interface AnimationsConfig {
  "column-switch"?: AnimationConfig | null;
  enabled?: boolean | null;
  overview?: AnimationConfig | null;
  "overview-window"?: AnimationConfig | null;
  slowdown?: number | null;
  "window-close"?: AnimationConfig | null;
  "window-move"?: AnimationConfig | null;
  "window-open"?: AnimationConfig | null;
  "window-resize"?: AnimationConfig | null;
  "workspace-switch"?: AnimationConfig | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "AnimationConfig".
 */
export interface AnimationConfig {
  duration?: number | null;
  easing?: EasingConfig | null;
  spring?: SpringConfig | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "EasingConfig".
 */
export interface EasingConfig {
  x1: number;
  x2: number;
  y1: number;
  y2: number;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "SpringConfig".
 */
export interface SpringConfig {
  damping?: number | null;
  epsilon?: number | null;
  mass?: number | null;
  stiffness?: number | null;
}